refactor(sbus): sbus refacotr
Took 2 minutes

+++ components/rc/include/channel_types.h
@@ -11,7 +11,7 @@ namespace rc
     /**
     * @brief Enum class representing the channel indices (1..16)
     */
-    enum class ChannelIndex: std::uint8_t {
+    enum class ChannelIndex: std::size_t {
         CH1 = 0, CH2, CH3, CH4, CH5, CH6, CH7, CH8,
         CH9, CH10, CH11, CH12, CH13, CH14, CH15, CH16,
         COUNT
@@ -23,5 +23,4 @@ namespace rc
     * @brief Scaled channel value that is expected to be within the range 0 - 2000.
     */
     using ChannelValue = int16_t;
-
 }
